Deciso_Making_IF_ELSE (Task 6-10)
TASK 6:
Write a program to generate the random number 1 to 6 (both inclusive, consider it a roll of dice) and print
English word (one to six) against the number?
Solution 6:
Using the rand function, generate random numbers and check if the number equals 1 then print the respective words using if else.

TASK 7:Write a program to roll dice two times for two players and print the number. Check and print whether player
1 is a winner or player 2 is a winner or a draw (the same number both times)?

Solution 7: Find the random number for  2 players 2 times and check which player's number is greater and then after checking using if condition and use counter to keep the count . 

Task 8: Write a program to roll dice six times and print number in each roll. Count and print, how many six
appeared in total?

Solution 8: Simply generate 6 random  numbers and then check if the number is six then count++ and at the end display the count.

Task 9: Write a program to ask user, which two numbers on dice are required. Roll the dice and print whether
user requirement is met?

Solution 9: Simply take the input from the user and then generate a random number twice. Match the input and random number and generate the output.

Task 10: Write a program for two players game. Roll dice 3 times for each player. Print numbers and their sum.
Finally, check and print if player 1 has won the game or player 2 has won the game or the game is drawn.

Solution 10: It's like creating a game but it's quite simple. Generate random numbers three times for each player and add all these 3 for player1 and player2 and then check which player's sum is greater . Print the output respectively.